Week 2 commit 2.
This commit changes the population
Separator file was changed.

populate_database file was changed.

Added read_database file, this file reads the data from the
MongoDB database and puts it in a data frame

Added split_data file, this file splits the data set being
used and prepares the data for oversampling. This is the
framework for oversampling. There are visualization tools to
view a scatter plot of the data set.