Week 3 commit 1.
This commit changes the split-data file

A number of functions were added that correspond to different
under and over sampling methods. These are used in the block
underneath the functions.

Also the plot_2d function was altered to operate in two
dimensions corresponding to occurence data that we are using
for the point observations.